feat: populate watermark settings to vertex spec
