+++ api/json-schema/schema.json
@@ -18559,7 +18559,7 @@
         },
         "watermark": {
           "$ref": "#/definitions/io.numaproj.numaflow.v1alpha1.Watermark",
-          "description": "Watermark enables watermark progression across the entire pipeline. Updating this after the pipeline has been created will have no impact and will be ignored. To make the pipeline honor any changes to the setting, the pipeline should be recreated."
+          "description": "Watermark enables watermark progression across the entire pipeline."
         }
       },
       "type": "object"
@@ -19018,6 +19018,10 @@
           "type": "array",
           "x-kubernetes-patch-merge-key": "name",
           "x-kubernetes-patch-strategy": "merge"
+        },
+        "watermark": {
+          "$ref": "#/definitions/io.numaproj.numaflow.v1alpha1.Watermark",
+          "description": "Watermark indicates watermark progression in the vertex, it's populated from the pipeline watermark settings."
         }
       },
       "required": [
