+++ pkg/apis/numaflow/v1alpha1/pipeline_types.go
@@ -371,9 +371,7 @@ type PipelineSpec struct {
 	// +kubebuilder:default={"readBatchSize": 500, "bufferMaxLength": 30000, "bufferUsageLimit": 80}
 	// +optional
 	Limits *PipelineLimits `json:"limits,omitempty" protobuf:"bytes,5,opt,name=limits"`
-	// Watermark enables watermark progression across the entire pipeline. Updating this after the pipeline has been
-	// created will have no impact and will be ignored. To make the pipeline honor any changes to the setting, the pipeline
-	// should be recreated.
+	// Watermark enables watermark progression across the entire pipeline.
 	// +kubebuilder:default={"disabled": false}
 	// +optional
 	Watermark Watermark `json:"watermark,omitempty" protobuf:"bytes,6,opt,name=watermark"`
