+++ pkg/apis/numaflow/v1alpha1/zz_generated.deepcopy.go
@@ -1734,6 +1734,7 @@ func (in *VertexSpec) DeepCopyInto(out *VertexSpec) {
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
+	in.Watermark.DeepCopyInto(&out.Watermark)
 	return
 }
