+++ pkg/reconciler/vertex/controller_test.go
@@ -199,8 +199,6 @@ func Test_BuildPodSpec(t *testing.T) {
 		assert.Contains(t, envNames, dfv1.EnvISBSvcSentinelMaster)
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisUser)
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisURL)
-		assert.Contains(t, envNames, dfv1.EnvWatermarkDisabled)
-		assert.Contains(t, envNames, dfv1.EnvWatermarkMaxDelay)
 		argStr := strings.Join(spec.InitContainers[0].Args, " ")
 		assert.Contains(t, argStr, "--buffers=")
 		for _, b := range testObj.GetToBuffers() {
@@ -237,8 +235,6 @@ func Test_BuildPodSpec(t *testing.T) {
 		assert.Contains(t, envNames, dfv1.EnvISBSvcSentinelMaster)
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisUser)
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisURL)
-		assert.Contains(t, envNames, dfv1.EnvWatermarkDisabled)
-		assert.Contains(t, envNames, dfv1.EnvWatermarkMaxDelay)
 		argStr := strings.Join(spec.InitContainers[0].Args, " ")
 		assert.Contains(t, argStr, "--buffers=")
 		for _, b := range testObj.GetFromBuffers() {
@@ -309,8 +305,6 @@ func Test_BuildPodSpec(t *testing.T) {
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisSentinelPassword)
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisUser)
 		assert.Contains(t, envNames, dfv1.EnvISBSvcRedisURL)
-		assert.Contains(t, envNames, dfv1.EnvWatermarkDisabled)
-		assert.Contains(t, envNames, dfv1.EnvWatermarkMaxDelay)
 		argStr := strings.Join(spec.InitContainers[0].Args, " ")
 		assert.Contains(t, argStr, "--buffers=")
 		for _, b := range testObj.GetFromBuffers() {
