--- pkg/shared/util/watermark.go
@@ -1,35 +1,0 @@
-/*
-Copyright 2022 The Numaproj Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import (
-	"time"
-
-	dfv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
-)
-
-// IsWatermarkEnabled returns true of watermark has been enabled.
-func IsWatermarkEnabled() bool {
-	return LookupEnvStringOr(dfv1.EnvWatermarkDisabled, "false") != "true"
-}
-
-// GetWatermarkMaxDelay returns user configured watermark max delay
-func GetWatermarkMaxDelay() time.Duration {
-	// The ENV variable is populated by the controller from validated user input, should not be invalid format.
-	d, _ := time.ParseDuration(LookupEnvStringOr(dfv1.EnvWatermarkMaxDelay, "0s"))
-	return d
-}
