+++ pkg/sources/generator/tickgen.go
@@ -22,6 +22,7 @@ import (
 	"context"
 	"encoding/binary"
 	"encoding/json"
+	"fmt"
 	"strconv"
 	"sync/atomic"
 	"time"
@@ -197,6 +198,13 @@ func NewMemGen(vertexInstance *dfv1.VertexInstance,
 	gensrc.forwarder = forwarder
 
 	return gensrc, nil
+}
+
+func (mg *memgen) buildSourceWatermarkPublisher(publishWMStores store.WatermarkStorer) publish.Publisher {
+	// for tickgen, it can be the name of the replica
+	entityName := fmt.Sprintf("%s-%d", mg.vertexInstance.Vertex.Name, mg.vertexInstance.Replica)
+	processorEntity := processor.NewProcessorEntity(entityName)
+	return publish.NewPublish(mg.lifecycleCtx, processorEntity, publishWMStores, publish.IsSource(), publish.WithDelay(mg.vertexInstance.Vertex.Spec.Watermark.GetMaxDelay()))
 }
 
 func (mg *memgen) GetName() string {
